+++ packages/flame/lib/src/components/position_component.dart
@@ -2,6 +2,7 @@ import 'dart:math' as math;
 import 'dart:ui' hide Offset;
 
 import '../anchor.dart';
+import '../effects/provider_interfaces.dart';
 import '../extensions/offset.dart';
 import '../extensions/vector2.dart';
 import '../game/notifying_vector2.dart';
@@ -57,18 +58,20 @@ import 'component.dart';
 /// the approximate bounding rectangle of the rendered picture. If you
 /// do not specify the size of a PositionComponent, then it will be
 /// equal to zero and the component won't be able to respond to taps.
-class PositionComponent extends Component {
+class PositionComponent extends Component
+    implements AngleProvider, PositionProvider, ScaleProvider {
   PositionComponent({
     Vector2? position,
     Vector2? size,
     Vector2? scale,
     double? angle,
     Anchor? anchor,
+    Iterable<Component>? children,
     int? priority,
   })  : transform = Transform2D(),
         _anchor = anchor ?? Anchor.topLeft,
         _size = NotifyingVector2.copy(size ?? Vector2.zero()),
-        super(priority: priority) {
+        super(children: children, priority: priority) {
     if (position != null) {
       transform.position = position;
     }
@@ -92,7 +95,9 @@ class PositionComponent extends Component {
   Matrix4 get transformMatrix => transform.transformMatrix;
 
   /// The position of this component's anchor on the screen.
+  @override
   NotifyingVector2 get position => transform.position;
+  @override
   set position(Vector2 position) => transform.position = position;
 
   /// X position of this component's anchor on the screen.
@@ -106,14 +111,18 @@ class PositionComponent extends Component {
   /// Rotation angle (in radians) of the component. The component will be
   /// rotated around its anchor point in the clockwise direction if the
   /// angle is positive, or counterclockwise if the angle is negative.
+  @override
   double get angle => transform.angle;
+  @override
   set angle(double a) => transform.angle = a;
 
   /// The scale factor of this component. The scale can be different along
   /// the X and Y dimensions. A scale greater than 1 makes the component
   /// bigger, and less than 1 smaller. The scale can also be negative,
   /// which results in a mirror reflection along the corresponding axis.
+  @override
   NotifyingVector2 get scale => transform.scale;
+  @override
   set scale(Vector2 scale) => transform.scale = scale;
 
   /// Anchor point for this component. An anchor point describes a point
@@ -178,17 +187,18 @@ class PositionComponent extends Component {
   /// has been applied.
   double get absoluteAngle {
     // TODO(spydon): take scale into consideration
-    return ancestors<PositionComponent>()
+    return ancestors()
+        .whereType<PositionComponent>()
         .fold<double>(angle, (totalAngle, c) => totalAngle + c.angle);
   }
 
   /// The resulting scale after all the ancestors and the components own scale
   /// has been applied.
   Vector2 get absoluteScale {
-    return ancestors<PositionComponent>().fold<Vector2>(
-      scale.clone(),
-      (totalScale, c) => totalScale..multiply(c.scale),
-    );
+    return ancestors().whereType<PositionComponent>().fold<Vector2>(
+          scale.clone(),
+          (totalScale, c) => totalScale..multiply(c.scale),
+        );
   }
 
   /// Measure the distance (in parent's coordinate space) between this