+++ packages/flame_forge2d/example/lib/domino_sample.dart
@@ -16,8 +16,7 @@ class Platform extends BodyComponent {
     final shape = PolygonShape()..setAsBoxXY(14.8, 0.125);
     final fixtureDef = FixtureDef(shape);
 
-    final bodyDef = BodyDef();
-    bodyDef.position = position;
+    final bodyDef = BodyDef(position: position);
     final body = world.createBody(bodyDef);
     return body..createFixture(fixtureDef);
   }
@@ -31,15 +30,14 @@ class DominoBrick extends BodyComponent {
   @override
   Body createBody() {
     final shape = PolygonShape()..setAsBoxXY(0.125, 2.0);
-    final fixtureDef = FixtureDef(shape)
-      ..density = 25.0
-      ..restitution = 0.4
-      ..friction = 0.5;
+    final fixtureDef = FixtureDef(
+      shape,
+      density: 25.0,
+      restitution: 0.4,
+      friction: 0.5,
+    );
 
-    final bodyDef = BodyDef()
-      ..type = BodyType.dynamic
-      ..position = position;
-
+    final bodyDef = BodyDef(type: BodyType.dynamic, position: position);
     return world.createBody(bodyDef)..createFixture(fixtureDef);
   }
 }
@@ -47,7 +45,7 @@ class DominoBrick extends BodyComponent {
 class DominoSample extends Forge2DGame with TapDetector {
   late Image pizzaImage;
 
-  DominoSample() : super(gravity: Vector2(0, -10.0));
+  DominoSample() : super(gravity: Vector2(0, 10.0));
 
   @override
   Future<void> onLoad() async {
@@ -55,17 +53,17 @@ class DominoSample extends Forge2DGame with TapDetector {
     boundaries.forEach(add);
     final center = screenToWorld(camera.viewport.effectiveSize / 2);
 
-    for (var i = 0; i < 8; i++) {
-      final position = center + Vector2(0.0, -30.0 + 5 * i);
+    const numberOfRows = 7;
+    for (var i = 0; i < numberOfRows - 2; i++) {
+      final position = center + Vector2(0.0, 5.0 * i);
       add(Platform(position));
     }
 
-    const numberOfRows = 10;
     const numberPerRow = 25;
     for (var i = 0; i < numberOfRows; ++i) {
       for (var j = 0; j < numberPerRow; j++) {
         final position = center +
-            Vector2(-14.75 + j * (29.5 / (numberPerRow - 1)), -27.7 + 5 * i);
+            Vector2(-14.75 + j * (29.5 / (numberPerRow - 1)), -12.7 + 5 * i);
         add(DominoBrick(position));
       }
     }
@@ -75,6 +73,6 @@ class DominoSample extends Forge2DGame with TapDetector {
   void onTapDown(TapDownInfo details) {
     super.onTapDown(details);
     final position = details.eventPosition.game;
-    add(Pizza(position));
+    add(Pizza(position)..renderBody = true);
   }
 }