+++ packages/flame_forge2d/lib/forge2d_game.dart
@@ -2,29 +2,24 @@ import 'package:flame/game.dart';
 import 'package:forge2d/forge2d.dart' hide Timer;
 
 import 'contact_callbacks.dart';
-import 'forge2d_camera.dart';
 
 class Forge2DGame extends FlameGame {
-  static final Vector2 defaultGravity = Vector2(0, -10.0);
-  static const double defaultZoom = 10.0;
-
-  final World world;
-
-  final ContactCallbacks _contactCallbacks = ContactCallbacks();
-
-  /// The camera translates the coordinate space after the viewport is applied.
-  @override
-  Forge2DCamera get camera => super.camera as Forge2DCamera;
-
   Forge2DGame({
     Vector2? gravity,
     double zoom = defaultZoom,
     Camera? camera,
   })  : world = World(gravity ?? defaultGravity),
-        super(camera: camera ?? Forge2DCamera()) {
+        super(camera: camera ?? Camera()) {
     this.camera.zoom = zoom;
     world.setContactListener(_contactCallbacks);
   }
+
+  static final Vector2 defaultGravity = Vector2(0, 10.0);
+  static const double defaultZoom = 10.0;
+
+  final World world;
+
+  final ContactCallbacks _contactCallbacks = ContactCallbacks();
 
   @override
   void update(double dt) {
@@ -50,5 +45,9 @@ class Forge2DGame extends FlameGame {
 
   Vector2 screenToWorld(Vector2 position) {
     return projector.unprojectVector(position);
+  }
+
+  Vector2 screenToFlameWorld(Vector2 position) {
+    return screenToWorld(position)..y *= -1;
   }
 }