+++ packages/flame_fire_atlas/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+ - Graduate package to a stable release. See pre-releases prior to this version for changelog entries.
+
+## 1.0.1-releasecandidate.1
+
 # CHANGELOG
 
 ## [1.0.0]