+++ packages/flame_tiled/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.0
+
+ - Graduate package to a stable release. See pre-releases prior to this version for changelog entries.
+
+## 1.3.0-releasecandidate.6
+
+ - Update a dependency to the latest release.
+
 ## 1.3.0-releasecandidate.5
 
  - Update a dependency to the latest release.